+++ core-ng/src/test/java/core/framework/impl/log/filter/MapParamTest.java
@@ -1,0 +1,20 @@
+package core.framework.impl.log.filter;
+
+import org.junit.jupiter.api.Test;
+
+import java.util.Map;
+import java.util.Set;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+/**
+ * @author neo
+ */
+class MapParamTest {
+    @Test
+    void filter() {
+        var param = new MapParam(Map.of("SessionId", "123", "key1", "value1"));
+        String message = param.filter(Set.of("SessionId"));
+        assertThat(message).contains("SessionId=******").contains("key1=value1");
+    }
+}